Way to obtaining float data from Google Analytics
Hello,




I work for a big marketing online company. We centralize our focus on web Analytics daily, and we use Gattica to obtain differents metrics that permits us to get data for our company. Using Gattica we found that the data that returns is integer and not float as we need to be more exactly obtaining metrics. We told with cannikin (Rob), answering if he could to say us the way to obtain this float data by changing the source code. He told us that we had to edit this file (data_point.rb) .
We investigate and now we can to obtain data with decimals, that permits us to get more exactly metrics.
This change is very importante to us, and we think that maybe can helps to another people that use this gem.
We propose this change for this file, or add the new file without deleting the original, because maybe some people only needs integer results and not float.




We hope that this helps the community, and wait your comments.




Thank you very much for your time.
Best Regards
Luna Mauro